Fixing Bugs and Updating InforPannel Page
