+++ src/config/IdeaPanel/IdeaPanel.json
